Check user access added to message methods

+++ src/messages/messages.service.ts
@@ -1,11 +1,13 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
+import { WsException } from '@nestjs/websockets';
 import { Model } from 'mongoose';
 import { Socket } from 'socket.io';
 import { Conversation, ConversationDocument } from '../conversations/conversation.schema';
 import { ConversationsService } from '../conversations/conversations.service';
 import { CreateMessageDto } from './dto/create-message.dto';
 import { UpdateMessageDto } from './dto/update-message.dto';
+import { LoggerWinston } from '../logger/logger-winston.service';
 import { Message, MessageDocument } from './message.schema';
 
 @Injectable()
@@ -14,9 +16,19 @@ export class MessagesService {
     @InjectModel(Conversation.name)
     private conversationModel: Model<ConversationDocument>,
     private readonly conversationsService: ConversationsService,
+    private readonly logger: LoggerWinston,
     @InjectModel(Message.name)
     private messageModel: Model<MessageDocument>,
   ) {}
+
+  private async checkUserAccess(id: string, client: Socket): Promise<void> {
+    const message = await this.messageModel.findById(id).exec();
+
+    if (message.user !== client.data.userID && message.conversation !== client.data.conversationID) {
+      this.logger.warn(`User error: forbidden access for user with id=${client.data.userID} to message with id=${id}`, 'MessagesService');
+      throw new WsException('Forbidden!');
+    }
+  }
 
   async create(createMessageDto: CreateMessageDto, client: Socket): Promise<Message> {
     const { text } = createMessageDto;
@@ -46,16 +58,20 @@ export class MessagesService {
     return this.messageModel.find({ conversation: conversationID }).exec();
   }
 
-  findOne(id: string): Promise<Message> {
+  async findOne(id: string, client: Socket): Promise<Message> {
+    await this.checkUserAccess(id, client);
     return this.messageModel.findById(id).exec();
   }
 
-  update(updateMessageDto: UpdateMessageDto): Promise<Message> {
+  async update(updateMessageDto: UpdateMessageDto, client: Socket): Promise<Message> {
     const { _id: id, ...updateMessageData } = updateMessageDto;
+    await this.checkUserAccess(id, client);
+
     return this.messageModel.findByIdAndUpdate(id, updateMessageData, { new: true }).exec();
   }
 
-  remove(id: string): Promise<Message> {
+  async remove(id: string, client: Socket): Promise<Message> {
+    await this.checkUserAccess(id, client);
     return this.messageModel.findByIdAndDelete(id).exec();
   }
 }
